Send email with new retrieved ads

This function is a wrapper of the emayili R package to send email messages with the information of the new ads

sendEmail(
  mailInfo = "_mailInfo.yml",
  newAds = new_KijijiAds(URL, excludeWords, pages, inputFile)
)

Arguments

mailInfo
character, the name of the file where all addresses and server information are saved

newAds
list, the output of the new_KijijiAds function

Details

The mailInfo file needs the following arguments:
from:
to:
host:
port:
username:
password:

Examples

if (FALSE) {
URL <- 'https://www.kijiji.ca/b-a-louer/ville-de-montreal/3-1-2/k0c30349001l1700281?price=__860'
sendEmail('_mailInfo.yml',
          newAds = new_KijijiAds(URL, excludeWords, pages = 2,
                                 inputFile = 'ads.json', updateInput = TRUE))
}